refactor(ui): make interface table columns reusable
Make some columns from the interface table reusable for all node types.

Fixes: canonical-web-and-design/app-tribe#570.
