+++ ui/src/app/base/components/node/networking/DHCPColumn/index.ts
