+++ ui/src/app/base/components/node/networking/FabricColumn/FabricColumn.tsx
@@ -5,10 +5,8 @@ import DoubleRow from "app/base/components/DoubleRow";
 import LegacyLink from "app/base/components/LegacyLink";
 import baseURLs from "app/base/urls";
 import fabricSelectors from "app/store/fabric/selectors";
-import machineSelectors from "app/store/machine/selectors";
-import type { Machine } from "app/store/machine/types";
 import type { RootState } from "app/store/root/types";
-import type { NetworkInterface, NetworkLink } from "app/store/types/node";
+import type { NetworkInterface, NetworkLink, Node } from "app/store/types/node";
 import { getInterfaceFabric, isBondOrBridgeParent } from "app/store/utils";
 import vlanSelectors from "app/store/vlan/selectors";
 import { getVLANDisplay } from "app/store/vlan/utils";
@@ -16,27 +14,24 @@ import { getVLANDisplay } from "app/store/vlan/utils";
 type Props = {
   link?: NetworkLink | null;
   nic?: NetworkInterface | null;
-  systemId: Machine["system_id"];
+  node: Node;
 };
 
-const FabricColumn = ({ link, nic, systemId }: Props): JSX.Element | null => {
-  const machine = useSelector((state: RootState) =>
-    machineSelectors.getById(state, systemId)
-  );
+const FabricColumn = ({ link, nic, node }: Props): JSX.Element | null => {
   const fabricsLoaded = useSelector(fabricSelectors.loaded);
   const fabrics = useSelector(fabricSelectors.all);
   const vlan = useSelector((state: RootState) =>
     vlanSelectors.getById(state, nic?.vlan_id)
   );
   const vlans = useSelector(vlanSelectors.all);
-  if (!machine) {
+  if (!node) {
     return null;
   }
   if (!fabricsLoaded) {
     return <Spinner />;
   }
-  const isABondOrBridgeParent = isBondOrBridgeParent(machine, nic, link);
-  const fabric = getInterfaceFabric(machine, fabrics, vlans, nic, link);
+  const isABondOrBridgeParent = isBondOrBridgeParent(node, nic, link);
+  const fabric = getInterfaceFabric(node, fabrics, vlans, nic, link);
   const fabricContent = !isABondOrBridgeParent
     ? fabric?.name || "Disconnected"
     : null;
