+++ ui/src/app/base/components/node/networking/FabricColumn/index.ts
