+++ ui/src/app/base/components/node/networking/NameColumn/index.ts
