+++ ui/src/app/base/components/node/networking/SubnetColumn/index.ts
