+++ ui/src/app/base/components/node/networking/TypeColumn/index.ts
