--- ui/src/app/machines/views/MachineDetails/MachineNetwork/NetworkTable/FabricColumn/FabricColumn.test.tsx
@@ -1,80 +1,0 @@
-import { mount } from "enzyme";
-import { Provider } from "react-redux";
-import configureStore from "redux-mock-store";
-
-import FabricColumn from "./FabricColumn";
-
-import type { RootState } from "app/store/root/types";
-import {
-  fabric as fabricFactory,
-  fabricState as fabricStateFactory,
-  machineDetails as machineDetailsFactory,
-  machineInterface as machineInterfaceFactory,
-  machineState as machineStateFactory,
-  machineStatus as machineStatusFactory,
-  rootState as rootStateFactory,
-  vlan as vlanFactory,
-} from "testing/factories";
-
-const mockStore = configureStore();
-
-describe("FabricColumn", () => {
-  let state: RootState;
-  beforeEach(() => {
-    state = rootStateFactory({
-      fabric: fabricStateFactory({
-        loaded: true,
-      }),
-      machine: machineStateFactory({
-        items: [machineDetailsFactory({ system_id: "abc123" })],
-        loaded: true,
-        statuses: {
-          abc123: machineStatusFactory(),
-        },
-      }),
-    });
-  });
-
-  it("displays a spinner if the data is loading", () => {
-    state.fabric.loaded = false;
-    const nic = machineInterfaceFactory();
-    state.machine.items = [
-      machineDetailsFactory({
-        interfaces: [nic],
-        system_id: "abc123",
-      }),
-    ];
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <FabricColumn nic={nic} systemId="abc123" />
-      </Provider>
-    );
-    expect(wrapper.find("DoubleRow").exists()).toBe(false);
-  });
-
-  it("can display fabric and vlan details", () => {
-    const fabric = fabricFactory({ name: "fabric-name" });
-    state.fabric.items = [fabric];
-    const vlan = vlanFactory({ fabric: fabric.id, vid: 2, name: "vlan-name" });
-    state.vlan.items = [vlan];
-    const nic = machineInterfaceFactory({
-      vlan_id: vlan.id,
-    });
-    state.machine.items = [
-      machineDetailsFactory({
-        interfaces: [nic],
-        system_id: "abc123",
-      }),
-    ];
-    const store = mockStore(state);
-    const wrapper = mount(
-      <Provider store={store}>
-        <FabricColumn nic={nic} systemId="abc123" />
-      </Provider>
-    );
-    const links = wrapper.find("DoubleRow LegacyLink");
-    expect(links.at(0).text()).toBe("fabric-name");
-    expect(links.at(1).text()).toBe("2 (vlan-name)");
-  });
-});
